# perf(types): remove redundant copy in match function