minor improvement on the efficiency of `.match`

master (https://github.com/jina-ai/jina/commit/63d59edca82003a07093016ee7c2aabfc925eac9):

```
➜  jina git:(master) ✗ python toy36.py
--- 12.913429542000001 seconds ---
➜  jina git:(master) ✗ python toy36.py
--- 13.194746499999999 seconds ---
➜  jina git:(master) ✗ python toy36.py
--- 14.238623624999999 seconds ---
```

this PR:

```
➜  jina git:(perf-minor-improve-match) ✗ python toy36.py
--- 6.968042625000001 seconds ---
➜  jina git:(perf-minor-improve-match) ✗ python toy36.py
--- 7.192200333000001 seconds ---
➜  jina git:(perf-minor-improve-match) ✗ python toy36.py
--- 6.7111269170000005 seconds ---
```

---

`toy36.py` grabbed from #3055 

```python
import time

import numpy as np

from jina import DocumentArray, Document


def main_func():
    da1 = DocumentArray()
    da2 = DocumentArray()
    for i in range(10_000):
        da1.append(Document(embedding=np.random.rand(512)))
        da2.append(Document(embedding=np.random.rand(512)))
    start_time = time.perf_counter()
    da1.match(da2, limit=10)
    print("--- %s seconds ---" % (time.perf_counter() - start_time))


main_func()

```